Updated to 2.0

+++ SimpleLogs4Net/Log.cs
@@ -1,81 +1,108 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.IO;
 namespace SimpleLogs4Net
 {
 	public class Log
 	{
-		private static string _Path;
-		private static bool _Enabled;
+		private static string _Dir;
+		public static bool _FileOutputEnabled;
+		public static bool _ConsoleOutputEnabled;
 		private static string _LastFile;
-		private static string _Prefix = "LOG";
+		public static string _Prefix = "LOG";
         private static int _Index = 0;
-        private static Event.Type _DefaultType = Event.Type.Normal;
-        public Log(string path, bool enabled, string prefix)
+        public static EType _DefaultType = EType.Normal;
+		private static void InitStream(OutputStream stream)
 		{
-			if (!Directory.Exists(path))
+			switch (stream)
+			{
+				case OutputStream.None:
+					_FileOutputEnabled = false;
+					_ConsoleOutputEnabled = false;
+					break;
+				case OutputStream.Console:
+					_FileOutputEnabled = false;
+					_ConsoleOutputEnabled = true;
+					break;
+				case OutputStream.File:
+					_FileOutputEnabled = true;
+					_ConsoleOutputEnabled = false;
+					break;
+				case OutputStream.Both:
+					_FileOutputEnabled = true;
+					_ConsoleOutputEnabled = true;
+					break;
+				default:
+					break;
+			}
+		}
+        public Log(string dir, OutputStream stream, string prefix)
+		{
+			if (!Directory.Exists(dir))
 			{
 				try
 				{
-					Directory.CreateDirectory(path);
+					Directory.CreateDirectory(dir);
 				}
 				catch
 				{
-					throw new FileNotFoundException("Directory Not Found", path);
+					throw new FileNotFoundException("Directory Not Found", dir);
 				}
 			}
-			_Path = path;
-			_Enabled = enabled;
+			_Dir = dir;
+			InitStream(stream);
 			_Prefix = prefix;
 		}
-		public Log(string path, bool enabled)
+		public Log(string dir, OutputStream stream)
 		{
-			if (!Directory.Exists(path))
+			if (!Directory.Exists(dir))
 			{
 				try
 				{
-					Directory.CreateDirectory(path);
+					Directory.CreateDirectory(dir);
 				}
 				catch
 				{
-					throw new FileNotFoundException("Directory Not Found", path);
+					throw new FileNotFoundException("Directory Not Found", dir);
 				}
 			}
-			_Path = path;
-			_Enabled = enabled;
+			_Dir = dir;
+			InitStream(stream);
 		}
-		public Log(string path,string prefix)
+		public Log(string dir,string prefix)
 		{
-			if (!Directory.Exists(path))
+			if (!Directory.Exists(dir))
 			{
 				try
 				{
-					Directory.CreateDirectory(path);
+					Directory.CreateDirectory(dir);
 				}
 				catch
 				{
-					throw new FileNotFoundException("Directory Not Found", path);
+					throw new FileNotFoundException("Directory Not Found", dir);
 				}
 				
 			}
-			_Path = path;
-			_Enabled = true;
+			_Dir = dir;
+			_FileOutputEnabled = true;
 			_Prefix = prefix;
 		}
-		public Log(string path)
+		public Log(string dir)
 		{
-			if (!Directory.Exists(path))
+			if (!Directory.Exists(dir))
 			{
 				try
 				{
-					Directory.CreateDirectory(path);
+					Directory.CreateDirectory(dir);
 				}
 				catch
 				{
-					throw new FileNotFoundException("Directory Not Found", path);
+					throw new FileNotFoundException("Directory Not Found", dir);
 				}
 			}
-			_Path = path;
-			_Enabled = true;
+			_Dir = dir;
+			_FileOutputEnabled = true;
 		}
         public static void Initialize(string path)
         {
@@ -85,27 +112,32 @@ namespace SimpleLogs4Net
         {
             new Log(path, prefix);
         }
-        public static void Initialize(string path, bool enabled)
+        public static void Initialize(string path, OutputStream stream)
         {
-            new Log(path, enabled);
+            new Log(path, stream);
         }
-        public static void Initialize(string path, bool enabled, string prefix)
+        public static void Initialize(string path, OutputStream stream, string prefix)
         {
-            new Log(path, enabled, prefix);
+            new Log(path, stream, prefix);
         }
-        public static void ChangeDefaultType(Event.Type type)
+		public static void NextLog() 
+		{
+			_LastFile = null;
+			_Index = 0;
+		}
+        public static void ChangeDefaultType(EType type)
         {
             _DefaultType = type;
         }
-        public static void ChangeEnable(bool enabled)
+        public static void ChangeOutputStream(OutputStream stream)
 		{
-			_Enabled = enabled;
+			InitStream(stream);
 		}
         public static void Write(string text)
         {
             AddEvent(new Event(text,_DefaultType));
         }
-		public static void Write(string text, Event.Type type)
+		public static void Write(string text, EType type)
 		{
 			AddEvent(new Event(text, type));
 		}
@@ -113,15 +145,73 @@ namespace SimpleLogs4Net
 		{
 			AddEvent(new Event(text, _DefaultType));
 		}
-		public static void Write(string[] text, Event.Type type)
+		public static void Write(string[] text, EType type)
 		{
 			AddEvent(new Event(text, type));
 		}
 		public static void AddEvent(Event logEvent)
 		{
-			if (!_Enabled)
+			if (!_FileOutputEnabled && !_ConsoleOutputEnabled)
 				return;
-			if (@_LastFile != null)
+            #region Console Output
+            if (_ConsoleOutputEnabled)
+			{
+				Console.ResetColor();
+				Console.Write("[");
+				Console.ForegroundColor = ConsoleColor.Cyan;
+				Console.Write(logEvent._DateTime.ToString("dd.MM.yyyy"));
+				Console.ResetColor();
+				Console.Write("-");
+				Console.ForegroundColor = ConsoleColor.Green;
+				Console.Write(logEvent._DateTime.ToString("HH:mm:ss"));
+				Console.ResetColor();
+				Console.Write("][");
+				switch (logEvent._Type)
+				{
+					default:
+					case EType.Normal:
+						Console.Write("NORMAL");
+						break;
+					case EType.Informtion:
+						Console.ForegroundColor = ConsoleColor.Blue;
+                        Console.Write("INFO");
+                        break;
+					case EType.Warning:
+                        Console.ForegroundColor = ConsoleColor.Yellow;
+                        Console.Write("WARNING");
+                        break;
+					case EType.Error:
+                        Console.ForegroundColor = ConsoleColor.DarkRed;
+                        Console.Write("ERROR");
+                        break;
+					case EType.Critical_Error:
+                        Console.ForegroundColor = ConsoleColor.Red;
+                        Console.Write("CRITICAL_ERROR");
+                        break;
+				}
+				Console.ResetColor();
+				Console.Write("]");
+				if (logEvent._IsMultiLine)
+				{
+					Console.WriteLine("[MULTILINE]");
+					Console.WriteLine("{");
+					foreach (string item in logEvent._MultiineText)
+					{
+						Console.WriteLine(item);
+					}
+					Console.WriteLine("}");
+                }
+				else
+				{
+					Console.WriteLine(logEvent._Text);
+				}
+				Console.ResetColor();
+			}
+#endregion
+            if (!_FileOutputEnabled)
+				return;
+            #region File Output
+            if (@_LastFile != null)
 			{
 				if (File.Exists(@_LastFile))
 				{
@@ -147,7 +237,7 @@ namespace SimpleLogs4Net
 					int i = 1;
 					do
 					{
-						_LastFile = _Path + _Prefix + i.ToString() + ".log";
+						_LastFile = _Dir + _Prefix + i.ToString() + ".log";
 						i++;
 					} while (File.Exists(@_LastFile));
 					StreamWriter writer = new StreamWriter(@_LastFile);
@@ -160,7 +250,7 @@ namespace SimpleLogs4Net
 				int i = 1;
 				do
 				{
-					_LastFile = _Path + _Prefix + i.ToString() + ".log";
+					_LastFile = _Dir + _Prefix + i.ToString() + ".log";
 					i++;
 				} while (File.Exists(@_LastFile));
 				StreamWriter writer = new StreamWriter(@_LastFile);
@@ -168,25 +258,26 @@ namespace SimpleLogs4Net
 				writer.Close();
 			}
             _Index = File.ReadAllLines(@_LastFile).Length;
+            #endregion
         }
-		public static string EventToString(Event logEvent)
+        public static string EventToString(Event logEvent)
 		{
 			string s = logEvent._DateTime.ToString("[dd.MM.yyyy-HH:mm:ss]");
 			switch (logEvent._Type)
 			{
-				case Event.Type.Normal:
+				case EType.Normal:
 					s += "[NORMAL]";
 					break;
-				case Event.Type.Informtion:
+				case EType.Informtion:
 					s += "[INFO]";
 					break;
-				case Event.Type.Warning:
+				case EType.Warning:
 					s += "[WARNING]";
 					break;
-				case Event.Type.Error:
+				case EType.Error:
 					s += "[ERROR]";
 					break;
-				case Event.Type.Critical_Error:
+				case EType.Critical_Error:
 					s += "[CRITICAL_ERROR]";
 					break;
 			}
@@ -212,19 +303,19 @@ namespace SimpleLogs4Net
 			string s = logEvent._DateTime.ToString("[dd.MM.yyyy-HH:mm:ss]");
 			switch (logEvent._Type)
 			{
-				case Event.Type.Normal:
+				case EType.Normal:
 					s += "[NORMAL]";
 					break;
-				case Event.Type.Informtion:
+				case EType.Informtion:
 					s += "[INFO]";
 					break;
-				case Event.Type.Warning:
+				case EType.Warning:
 					s += "[WARNING]";
 					break;
-				case Event.Type.Error:
+				case EType.Error:
 					s += "[ERROR]";
 					break;
-				case Event.Type.Critical_Error:
+				case EType.Critical_Error:
 					s += "[CRITICAL_ERROR]";
 					break;
 			}
@@ -247,22 +338,13 @@ namespace SimpleLogs4Net
 		}
 		public static void ClearLogs()
 		{
-			List<string> LogFiles = new List<string>();
-			string x;
-			int i = 1;
-			do
-			{
-				x = _Path + _Prefix + i.ToString() + ".log";
-				if (File.Exists(@x))
-				{
-					LogFiles.Add(x);
-				}
-				i++;
-			} while (File.Exists(@x));
-			foreach (string item in LogFiles)
-			{
-				File.Delete(item);
-			}
+            foreach (string item in Directory.GetFiles(_Dir))
+            {
+                if (item.EndsWith(".log"))
+                {
+					File.Delete(item);
+                }
+            }
 		}
 	}
 }
